requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
numpy = "*"
pandas = "*"
seaborn = "*"
pandas-profiling = "*"
ipywidgets = "*"
sklearn = "*"
shap = "*"
catboost = "*"
kafka-python = "*"
libcst = "*"
python-dotenv = "*"
elasticsearch = "*"
aiohttp = "==3.7.2"